fix: make jsd_mp a regular package (add __init__) and fix cache typing
- src/jsd_mp had no top-level __init__.py, so it resolved as an implicit
  namespace package. That worked locally but broke in CI, where subpackage
  imports (jsd_mp.config) failed intermittently depending on sys.path order.
  Add src/jsd_mp/__init__.py so it is a proper regular package.
- with the package now resolving correctly, mypy flagged Bari.get_cost's
  struct_path_len cache (typed float, holding int, .get -> int|None). Type the
  cache as int and narrow via a local before assignment.

+++ src/jsd_mp/solver/solver.py
@@ -39,7 +39,7 @@ class Solver(abc.ABC):
         # between node pairs. The topology's connections never change during a
         # solve, so these are constant and worth memoizing across the hot loop.
         self.struct_path_len: typing.Dict[
-            typing.Tuple[str, str], float
+            typing.Tuple[str, str], int
         ] = {}
 
         self.solved: bool = False
